+++ lib/active_job/queue_adapters/sidekiq_adapter.rb
@@ -16,7 +16,7 @@ end
 module ActiveJob
   module QueueAdapters
     # Explicitly remove the implementation existing in older rails'.
-    remove_const(:SidekiqAdapter) if defined?(:SidekiqAdapter)
+    remove_const(:SidekiqAdapter) if const_defined?(:SidekiqAdapter)
 
     # Sidekiq adapter for Active Job
     #