+++ myapp/config/routes.rb
@@ -2,8 +2,11 @@
 ENV["SIDEKIQ_WEB_TESTING"] = "1"
 
 require "sidekiq/web"
+Sidekiq::Web.configure do |config|
+  config.app_url = "/"
+  # config.custom_job_info_rows << RowHelper.new
+end
 require "sidekiq-redis_info/web"
-Sidekiq::Web.app_url = "/"
 
 Rails.application.routes.draw do
   mount Sidekiq::Web => "/sidekiq"